Nunca dibuje lo que pueda copiar. Nunca copie lo que pueda calcar. Nunca calque lo que pueda recortar y pegar. 		-- Regla de Wallace.

+++ provision/dotnet/post_install.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
-cowsay "Start Post Install Script"
-cp -r /usr/share/dotnet/sdk/ /usr/lib64/dotnet/
-dotnet tool install --global dotnet-ef --version 5.0.10
-cowsay "End Post Install Script"
+FILE_CHECK=".copy_paste_dotnet_sdk"
+
+if [ ! -f ~/$FILE_CHECK ]
+then
+	cowsay "Start Post Install Script"
+	cp -rv /usr/share/dotnet/sdk/ /usr/lib64/dotnet/
+	dotnet tool install --global dotnet-ef --version 5.0.10
+	cowsay "End Post Install Script"
+fi
